+++ tests/integration/tx/aminojson/aminojson_test.go
@@ -1,7 +1,12 @@
 package aminojson
 
 import (
+<<<<<<< HEAD
 	"context"
+=======
+	"bytes"
+	"encoding/json"
+>>>>>>> 60cbe2db2 (fix(x/tx): sort with oneof field name in amino-json (#21782))
 	"fmt"
 	"reflect"
 	"testing"
@@ -572,4 +577,21 @@ func postFixPulsarMessage(msg proto.Message) {
 			m.Permissions = nil
 		}
 	}
+}
+
+// sortJson sorts the JSON bytes by way of the side effect of unmarshalling and remarshalling the JSON
+// using encoding/json.  This hacky way of sorting JSON fields was used by the legacy amino JSON encoding in
+// x/auth/migrations/legacytx.StdSignBytes.  It is used here ensure the x/tx JSON encoding is equivalent to
+// the legacy amino JSON encoding.
+func sortJson(bz []byte) ([]byte, error) {
+	var c any
+	err := json.Unmarshal(bz, &c)
+	if err != nil {
+		return nil, err
+	}
+	js, err := json.Marshal(c)
+	if err != nil {
+		return nil, err
+	}
+	return js, nil
 }